ootb: refer to instructions for private git repo
in order for the workload to work with a private git repository, it's a
must that a corev1/Secret object exist in the same namespace as the
Workload (such that `gitrepository.spec.secretRef: {name: <>}` points at
it).

while this has been part of the `scc/*` documentation, it's been lacking
from the `getting-started`, making folks that configure private
repositories to be unable to have their source code being properly
fetched.

the tl;dr of the instructions is: want to use private git source?

1. in the dev namespace, create a secret w/ git credentials
2. update the workload with `--param=source_git_ssh_secret=developer-git-credentials`

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>